# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Sarathi API
  version: 0.1.0
paths:
...
  /api/v1/lessons/{lesson_id}/checkpoint:
    post:
      tags:
      - Checkpoints
      summary: Generate Checkpoint
      operationId: generate_checkpoint_api_v1_lessons__lesson_id__checkpoint_post
      security:
      - HTTPBearer: []
      parameters:
      - name: lesson_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lesson Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/checkpoints/{checkpoint_id}/submit:
    post:
      tags:
      - Checkpoints
      summary: Submit Checkpoint
      operationId: submit_checkpoint_api_v1_checkpoints__checkpoint_id__submit_post
      security:
      - HTTPBearer: []
      parameters:
      - name: checkpoint_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Checkpoint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckpointSubmitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointAttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/goals:
    post:
      tags:
      - Goals
      summary: Create Goal
      operationId: create_goal_api_v1_goals_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Goals
      summary: List Goals
      operationId: list_goals_api_v1_goals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Size
      - name: sort_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort By
      - name: sort_desc
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Sort Desc
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_GoalResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/goals/{goal_id}:
    patch:
      tags:
      - Goals
      summary: Update Goal
      operationId: update_goal_api_v1_goals__goal_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: goal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Goal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/goals/{goal_id}/plan:
    post:
      tags:
      - Goals
      summary: Generate Plan Endpoint
      operationId: generate_plan_endpoint_api_v1_goals__goal_id__plan_post
      security:
      - HTTPBearer: []
      parameters:
      - name: goal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Goal Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CheckpointAttemptResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        score:
          type: number
          title: Score
        mastery_deltas:
          items:
            $ref: '#/components/schemas/MasteryDelta'
          type: array
          title: Mastery Deltas
        feedback:
          items:
            $ref: '#/components/schemas/ItemFeedback'
          type: array
          title: Feedback
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
      type: object
      required:
      - id
      - score
      - mastery_deltas
      - feedback
      - submitted_at
      title: CheckpointAttemptResponse
    CheckpointItem:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          title: Type
        question:
          type: string
          title: Question
        options:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Options
        concept_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Concept Ids
      type: object
      required:
      - id
      - type
      - question
      - concept_ids
      title: CheckpointItem
    CheckpointResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        lesson_id:
          type: string
          format: uuid
          title: Lesson Id
        items:
          items:
            $ref: '#/components/schemas/CheckpointItem'
          type: array
          title: Items
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - lesson_id
      - items
      - created_at
      title: CheckpointResponse
    CheckpointSubmitRequest:
      properties:
        responses:
          additionalProperties: true
          type: object
          title: Responses
      type: object
      required:
      - responses
      title: CheckpointSubmitRequest
...
    GoalCreate:
      properties:
        raw_input:
          type: string
          minLength: 10
          title: Raw Input
          description: The raw natural language input describing the goal.
      type: object
      required:
      - raw_input
      title: GoalCreate
    GoalResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        raw_input:
          type: string
          title: Raw Input
        normalized_topic:
          type: string
          title: Normalized Topic
        target_level:
          type: string
          title: Target Level
        deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Deadline
        motivation_hint:
          anyOf:
          - type: string
          - type: 'null'
          title: Motivation Hint
        is_educational:
          type: boolean
          title: Is Educational
        clarification_needed:
          anyOf:
          - type: string
          - type: 'null'
          title: Clarification Needed
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - raw_input
      - normalized_topic
      - target_level
      - deadline
      - motivation_hint
      - is_educational
      - clarification_needed
      - status
      - created_at
      - updated_at
      title: GoalResponse
    GoalUpdate:
      properties:
        normalized_topic:
          anyOf:
          - type: string
          - type: 'null'
          title: Normalized Topic
          description: Corrected topic.
        target_level:
          anyOf:
          - type: string
            enum:
            - beginner
            - intermediate
            - advanced
          - type: 'null'
          title: Target Level
          description: Corrected target level.
        deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Deadline
          description: Corrected deadline.
      type: object
      title: GoalUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ItemFeedback:
      properties:
        item_id:
          type: string
          title: Item Id
        correct:
          type: boolean
          title: Correct
        explanation:
          type: string
          title: Explanation
      type: object
      required:
      - item_id
      - correct
      - explanation
      title: ItemFeedback
...
    MasteryDelta:
      properties:
        concept_id:
          type: string
          format: uuid
          title: Concept Id
        delta:
          type: number
          title: Delta
      type: object
      required:
      - concept_id
      - delta
      title: MasteryDelta
...
    PaginatedResponse_GoalResponse_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/GoalResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - data
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[GoalResponse]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer